# api/routes/web.php
<?php

use App\Http\Controllers\Api\ImportExportController;
use App\Http\Controllers\Api\ItemImageController;
use App\Http\Controllers\Api\ItemStatusController;
use App\Http\Controllers\Api\PurchaseCandidateController;
use App\Http\Controllers\Api\SettingsTpoController;
use App\Http\Controllers\Api\WearLogController;
use App\Http\Controllers\Api\WeatherLocationController;
use App\Http\Controllers\Api\WeatherRecordController;
use App\Http\Controllers\AuthController;
use App\Http\Requests\ItemStoreRequest;
use App\Http\Requests\ItemUpdateRequest;
use App\Models\CategoryMaster;
use App\Models\Item;
use App\Models\Outfit;
use App\Services\Brands\UserBrandService;
use App\Services\Items\ItemStoreService;
use App\Services\Items\ItemUpdateService;
use App\Services\Settings\UserPreferenceService;
use App\Services\Settings\UserTpoService;
use App\Support\DisposedItemsIndexQuery;
use App\Support\HomeSummaryQuery;
use App\Support\ItemPayloadBuilder;
use App\Support\ItemsIndexQuery;
use App\Support\ItemSubcategorySupport;
use App\Support\OutfitPayloadBuilder;
use App\Support\OutfitsIndexQuery;
use App\Support\SkinTonePresetSupport;
use App\Support\TpoSelectionResolver;
use App\Support\UserPreferencePayloadBuilder;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\DB;
use Illuminate\Support\Facades\Route;

if (! function_exists('defaultRolledOutVisibleCategoryIds')) {
    function defaultRolledOutVisibleCategoryIds(): array
    {
        return [
            'underwear_bra',
            'underwear_shorts',
            'underwear_shapewear',
            'underwear_undershirt',
            'underwear_other',
        ];
    }
}

if (! function_exists('resolveVisibleCategoryIdsForSettings')) {
    function resolveVisibleCategoryIdsForSettings($user, array $activeCurrentVisibleCategoryIds): array
    {
        $visibleCategoryIds = $user->visible_category_ids;

        if (! is_array($visibleCategoryIds)) {
            return $activeCurrentVisibleCategoryIds;
        }

        $normalizedVisibleCategoryIds = array_values(array_filter(
            $visibleCategoryIds,
            fn ($id) => is_string($id) && in_array($id, $activeCurrentVisibleCategoryIds, true)
        ));

        foreach ($activeCurrentVisibleCategoryIds as $categoryId) {
            if (
                in_array($categoryId, defaultRolledOutVisibleCategoryIds(), true)
                && ! in_array($categoryId, $normalizedVisibleCategoryIds, true)
            ) {
                $normalizedVisibleCategoryIds[] = $categoryId;
            }
        }

        return $normalizedVisibleCategoryIds;
    }
}

Route::get('/csrf-cookie', function (Request $request) {
    // セッション開始＆CSRFトークン生成
    $request->session()->start();
    $request->session()->regenerateToken();

    $token = $request->session()->token();

    return response()->noContent()->cookie(
        'XSRF-TOKEN',
        $token,
        120,   // minutes
        '/',
        null,
        false, // secure（ローカルはfalse）
        false, // httpOnly（JS/クライアントが読む想定）
        false,
        'Lax'
    );
});

Route::prefix('api')->middleware(['web'])->group(function () {
    // Auth
    Route::post('/register', [AuthController::class, 'register']);
    Route::post('/login', [AuthController::class, 'login']);

    Route::middleware('auth:web')->group(function () {
        Route::get('/me', function (Request $request) {
            $user = $request->user();

            return response()->json([
                'id' => $user->id,
                'name' => $user->name,
                'email' => $user->email,
            ]);
        });

        Route::post('/logout', [AuthController::class, 'logout']);
    });

    Route::middleware('auth:web')->get('/home/summary', function (Request $request) {
        return response()->json(HomeSummaryQuery::build($request->user()));
    });

    Route::middleware('auth:web')->controller(ImportExportController::class)->group(function () {
        Route::get('/export', 'export');
        Route::post('/import', 'import');
    });

    // Settings
    Route::middleware('auth:web')->get('/settings/categories', function (Request $request) {
        $user = $request->user();
        $activeCurrentVisibleCategoryIds = CategoryMaster::query()
            ->where('is_active', true)
            ->whereIn('id', ItemSubcategorySupport::currentVisibleCategoryIds())
            ->orderBy('group_id')
            ->orderBy('sort_order')
            ->pluck('id')
            ->all();

        $visibleCategoryIds = resolveVisibleCategoryIdsForSettings(
            $user,
            $activeCurrentVisibleCategoryIds,
        );

        return response()->json([
            'visibleCategoryIds' => $visibleCategoryIds,
        ]);
    });

    Route::middleware('auth:web')->put('/settings/categories', function (Request $request) {
        $validated = $request->validate([
            'visibleCategoryIds' => ['present', 'array'],
            'visibleCategoryIds.*' => ['string', 'distinct'],
        ]);

        $activeCurrentVisibleCategoryIds = CategoryMaster::query()
            ->where('is_active', true)
            ->whereIn('id', ItemSubcategorySupport::currentVisibleCategoryIds())
            ->pluck('id')
            ->all();

        $visibleCategoryIds = collect($validated['visibleCategoryIds'])
            ->values();

        $activeCount = $visibleCategoryIds
            ->filter(fn ($id) => in_array($id, $activeCurrentVisibleCategoryIds, true))
            ->count();

        if ($activeCount !== $visibleCategoryIds->count()) {
            return response()->json([
                'message' => '無効なカテゴリが含まれています。',
            ], 422);
        }

        $user = $request->user();
        $user->forceFill([
            'visible_category_ids' => $visibleCategoryIds->all(),
        ])->save();

        return response()->json([
            'message' => 'updated',
            'visibleCategoryIds' => $user->visible_category_ids ?? [],
        ]);
    });

    Route::middleware('auth:web')->get('/settings/brands', function (Request $request) {
        $validated = $request->validate([
            'active_only' => ['nullable', 'boolean'],
            'keyword' => ['nullable', 'string', 'max:255'],
        ]);

        $brands = app(UserBrandService::class)->list($request->user(), $validated);

        return response()->json([
            'brands' => collect($brands)->map(fn ($brand) => [
                'id' => $brand->id,
                'name' => $brand->name,
                'kana' => $brand->kana,
                'is_active' => $brand->is_active,
                'updated_at' => optional($brand->updated_at)?->toISOString(),
            ])->all(),
        ]);
    });

    Route::middleware('auth:web')->post('/settings/brands', function (Request $request) {
        $validated = $request->validate([
            'name' => ['required', 'string', 'max:255'],
            'kana' => ['nullable', 'string', 'max:255'],
            'is_active' => ['nullable', 'boolean'],
        ], [
            'name.required' => 'ブランド名を入力してください。',
        ]);

        $brand = app(UserBrandService::class)->create($request->user(), $validated);

        return response()->json([
            'message' => 'created',
            'brand' => [
                'id' => $brand->id,
                'name' => $brand->name,
                'kana' => $brand->kana,
                'is_active' => $brand->is_active,
                'updated_at' => optional($brand->updated_at)?->toISOString(),
            ],
        ], 201);
    });

    Route::middleware('auth:web')->patch('/settings/brands/{id}', function (Request $request, int $id) {
        $validated = $request->validate([
            'name' => ['sometimes', 'required', 'string', 'max:255'],
            'kana' => ['sometimes', 'nullable', 'string', 'max:255'],
            'is_active' => ['sometimes', 'boolean'],
        ], [
            'name.required' => 'ブランド名を入力してください。',
        ]);

        if ($validated === []) {
            return response()->json([
                'message' => '更新項目がありません。',
            ], 422);
        }

        $brand = app(UserBrandService::class)->update($request->user(), $id, $validated);

        return response()->json([
            'message' => 'updated',
            'brand' => [
                'id' => $brand->id,
                'name' => $brand->name,
                'kana' => $brand->kana,
                'is_active' => $brand->is_active,
                'updated_at' => optional($brand->updated_at)?->toISOString(),
            ],
        ]);
    });

    Route::middleware('auth:web')->get('/settings/preferences', function (Request $request) {
        $preference = app(UserPreferenceService::class)->get($request->user());

        return response()->json([
            'preferences' => UserPreferencePayloadBuilder::build($preference),
        ]);
    });

    Route::middleware('auth:web')->put('/settings/preferences', function (Request $request) {
        $validated = $request->validate([
            'currentSeason' => ['nullable', 'string', 'in:spring,summer,autumn,winter'],
            'defaultWearLogStatus' => ['nullable', 'string', 'in:planned,worn'],
            'calendarWeekStart' => ['nullable', 'string', 'in:monday,sunday'],
            'skinTonePreset' => ['nullable', 'string', 'in:'.implode(',', SkinTonePresetSupport::values())],
        ]);

        $preference = app(UserPreferenceService::class)->update($request->user(), $validated);

        return response()->json([
            'message' => 'updated',
            'preferences' => UserPreferencePayloadBuilder::build($preference),
        ]);
    });

    Route::middleware('auth:web')->controller(SettingsTpoController::class)->group(function () {
        Route::get('/settings/tpos', 'index');
        Route::post('/settings/tpos', 'store');
        Route::patch('/settings/tpos/{id}', 'update');
    });

    Route::middleware('auth:web')->controller(WeatherLocationController::class)->group(function () {
        Route::get('/settings/weather-locations', 'index');
        Route::post('/settings/weather-locations', 'store');
        Route::patch('/settings/weather-locations/{id}', 'update');
        Route::delete('/settings/weather-locations/{id}', 'destroy');
    });

    Route::middleware('auth:web')->controller(WeatherRecordController::class)->group(function () {
        Route::get('/weather-records', 'index');
        Route::post('/weather-records', 'store');
        Route::patch('/weather-records/{id}', 'update');
        Route::delete('/weather-records/{id}', 'destroy');
    });

    // Items
    Route::middleware('auth:web')->get('/items', function (Request $request) {
        return response()->json(ItemsIndexQuery::build($request->user(), $request));
    });

    Route::middleware('auth:web')->get('/items/disposed', function (Request $request) {
        return response()->json(DisposedItemsIndexQuery::build($request->user(), $request));
    });

    Route::middleware('auth:web')->post('/items', function (ItemStoreRequest $request) {
        $validated = $request->validated();

        $item = app(ItemStoreService::class)->store($request->user(), $validated);

        return response()->json([
            'message' => 'created',
            'item' => ItemPayloadBuilder::buildDetail($item),
        ], 201);
    });

    Route::middleware('auth:web')->get('/items/{id}', function (Request $request, int $id) {
        $item = Item::query()
            ->where('user_id', $request->user()->id)
            ->with(['images', 'materials', 'user'])
            ->findOrFail($id);

        return response()->json([
            'item' => ItemPayloadBuilder::buildDetail($item),
        ]);
    });

    Route::middleware('auth:web')->put('/items/{id}', function (ItemUpdateRequest $request, int $id) {
        $item = Item::query()
            ->where('user_id', $request->user()->id)
            ->findOrFail($id);
        $validated = $request->validated();

        $item = app(ItemUpdateService::class)->update($request->user(), $item, $validated);

        return response()->json([
            'message' => 'updated',
            'item' => ItemPayloadBuilder::buildDetail($item),
        ]);
    });

    Route::middleware('auth:web')->controller(ItemStatusController::class)->group(function () {
        Route::post('/items/{id}/dispose', 'dispose');
        Route::post('/items/{id}/reactivate', 'reactivate');
        Route::post('/items/{id}/care-status', 'updateCareStatus');
    });

    Route::middleware('auth:web')->delete('/items/{id}', function (Request $request, int $id) {
        $item = Item::query()
            ->where('user_id', $request->user()->id)
            ->findOrFail($id);

        $hasOutfitReference = $item->outfitItems()->exists();
        $hasWearLogReference = $item->wearLogItems()->exists();

        if ($hasOutfitReference || $hasWearLogReference) {
            return response()->json([
                'message' => 'このアイテムは参照中のため完全に削除できません。手放す操作を利用してください。',
            ], 422);
        }

        $item->delete();

        return response()->json([
            'message' => 'deleted',
        ]);
    });

    Route::middleware('auth:web')->controller(ItemImageController::class)->group(function () {
        Route::post('/items/{id}/images', 'store');
        Route::delete('/items/{id}/images/{imageId}', 'destroy');
    });

    // Purchase Candidates
    Route::middleware('auth:web')->controller(PurchaseCandidateController::class)->group(function () {
        Route::get('/purchase-candidates', 'index');
        Route::post('/purchase-candidates', 'store');
        Route::get('/purchase-candidates/{id}', 'show');
        Route::put('/purchase-candidates/{id}', 'update');
        Route::delete('/purchase-candidates/{id}', 'destroy');
        Route::post('/purchase-candidates/{id}/duplicate', 'duplicate');
        Route::post('/purchase-candidates/{id}/color-variant', 'colorVariant');
        Route::post('/purchase-candidates/{id}/images', 'uploadImage');
        Route::delete('/purchase-candidates/{id}/images/{imageId}', 'deleteImage');
        Route::post('/purchase-candidates/{id}/item-draft', 'itemDraft');
    });

    // Outfits
    Route::middleware('auth:web')->get('/outfits', function (Request $request) {
        return response()->json(OutfitsIndexQuery::build($request->user(), $request));
    });

    Route::middleware('auth:web')->get('/outfits/invalid', function (Request $request) {
        return response()->json(OutfitsIndexQuery::build($request->user(), $request, 'invalid'));
    });

    Route::middleware('auth:web')->post('/outfits/{id}/restore', function (Request $request, int $id) {
        $outfit = Outfit::query()
            ->where('user_id', $request->user()->id)
            ->with(['outfitItems.item'])
            ->findOrFail($id);

        if ($outfit->status !== 'invalid') {
            return response()->json([
                'message' => 'このコーディネートは復帰できません。',
            ], 422);
        }

        $hasInactiveItems = $outfit->outfitItems->contains(function ($outfitItem) {
            return $outfitItem->item === null || $outfitItem->item->status !== 'active';
        });

        if ($hasInactiveItems) {
            return response()->json([
                'message' => 'このコーディネートは復帰できません。',
            ], 422);
        }

        $outfit->update([
            'status' => 'active',
        ]);

        return response()->json([
            'message' => 'restored',
            'outfit' => OutfitPayloadBuilder::buildDetail(
                $outfit->fresh()->load(['outfitItems.item', 'user'])
            ),
        ]);
    });

    Route::middleware('auth:web')->post('/outfits/{id}/duplicate', function (Request $request, int $id) {
        $outfit = Outfit::query()
            ->where('user_id', $request->user()->id)
            ->with(['outfitItems.item', 'user'])
            ->findOrFail($id);

        $payloadItems = $outfit->outfitItems
            ->sortBy('sort_order')
            ->values()
            ->map(function ($outfitItem) use ($outfit) {
                $isDisposed = $outfit->status === 'invalid'
                    && ($outfitItem->item === null || $outfitItem->item->status !== 'active');

                return [
                    'item_id' => $outfitItem->item_id,
                    'sort_order' => $outfitItem->sort_order,
                    'selectable' => ! $isDisposed,
                    'note' => $isDisposed ? '手放したアイテムのため初期選択から除外' : null,
                ];
            })
            ->all();

        $duplicatedName = $outfit->name !== null
            ? $outfit->name.'（コピー）'
            : '（コピー）';
        $outfitPayload = OutfitPayloadBuilder::buildDetail($outfit);

        return response()->json([
            'message' => 'duplicated_payload_ready',
            'outfit' => [
                'name' => $duplicatedName,
                'memo' => $outfit->memo,
                'seasons' => $outfit->seasons ?? [],
                'tpos' => $outfitPayload['tpos'],
                'tpo_ids' => $outfitPayload['tpo_ids'],
                'items' => $payloadItems,
            ],
        ]);
    });

    Route::middleware('auth:web')->post('/outfits', function (Request $request) {
        $validated = $request->validate([
            'name' => ['nullable', 'string', 'max:255'],
            'memo' => ['nullable', 'string'],
            'seasons' => ['nullable', 'array'],
            'seasons.*' => ['string', 'max:50'],
            'tpo_ids' => ['nullable', 'array'],
            'tpo_ids.*' => ['integer'],
            'tpos' => ['nullable', 'array'],
            'tpos.*' => ['string', 'max:50'],
            'items' => ['required', 'array', 'min:1'],
            'items.*.item_id' => ['required', 'integer'],
            'items.*.sort_order' => ['required', 'integer', 'min:1'],
        ]);

        $user = $request->user();

        $itemIds = collect($validated['items'])
            ->pluck('item_id')
            ->unique()
            ->values();

        $availableItemCount = Item::query()
            ->where('user_id', $user->id)
            ->where('status', 'active')
            ->whereIn('id', $itemIds)
            ->count();

        // 他人のitem_idや disposed item は使用不可
        if ($availableItemCount !== $itemIds->count()) {
            return response()->json([
                'message' => '選択したアイテムに不正なデータが含まれています。',
            ], 422);
        }

        $outfit = DB::transaction(function () use ($validated, $user) {
            $outfit = Outfit::create([
                'user_id' => $user->id,
                'name' => $validated['name'] ?? null,
                'memo' => $validated['memo'] ?? null,
                'seasons' => $validated['seasons'] ?? [],
                'tpo_ids' => TpoSelectionResolver::resolve(app(UserTpoService::class), $user, $validated),
            ]);

            $outfit->outfitItems()->createMany(
                collect($validated['items'])->map(function ($item) {
                    return [
                        'item_id' => $item['item_id'],
                        'sort_order' => $item['sort_order'],
                    ];
                })->all()
            );

            return $outfit->load(['outfitItems.item', 'user']);
        });

        return response()->json([
            'message' => 'created',
            'outfit' => OutfitPayloadBuilder::buildDetail($outfit),
        ], 201);
    });

    Route::middleware('auth:web')->get('/outfits/{id}', function (Request $request, int $id) {
        $outfit = Outfit::query()
            ->where('user_id', $request->user()->id)
            ->with(['outfitItems.item', 'user'])
            ->findOrFail($id);

        return response()->json([
            'outfit' => OutfitPayloadBuilder::buildDetail($outfit),
        ]);
    });

    Route::middleware('auth:web')->put('/outfits/{id}', function (Request $request, int $id) {
        $outfit = Outfit::query()
            ->where('user_id', $request->user()->id)
            ->findOrFail($id);

        $validated = $request->validate([
            'name' => ['nullable', 'string', 'max:255'],
            'memo' => ['nullable', 'string'],
            'seasons' => ['nullable', 'array'],
            'seasons.*' => ['string', 'max:50'],
            'tpo_ids' => ['nullable', 'array'],
            'tpo_ids.*' => ['integer'],
            'tpos' => ['nullable', 'array'],
            'tpos.*' => ['string', 'max:50'],
            'items' => ['required', 'array', 'min:1'],
            'items.*.item_id' => ['required', 'integer'],
            'items.*.sort_order' => ['required', 'integer', 'min:1'],
        ]);

        $user = $request->user();

        $itemIds = collect($validated['items'])
            ->pluck('item_id')
            ->unique()
            ->values();

        $availableItemCount = Item::query()
            ->where('user_id', $user->id)
            ->where('status', 'active')
            ->whereIn('id', $itemIds)
            ->count();

        if ($availableItemCount !== $itemIds->count()) {
            return response()->json([
                'message' => '選択したアイテムに不正なデータが含まれています。',
            ], 422);
        }

        DB::transaction(function () use ($outfit, $validated, $user) {
            $outfit->update([
                'name' => $validated['name'] ?? null,
                'memo' => $validated['memo'] ?? null,
                'seasons' => $validated['seasons'] ?? [],
                'tpo_ids' => TpoSelectionResolver::resolve(
                    app(UserTpoService::class),
                    $user,
                    $validated,
                    $outfit->tpo_ids ?? [],
                ),
            ]);

            $outfit->outfitItems()->delete();

            $outfit->outfitItems()->createMany(
                collect($validated['items'])->map(function ($item) {
                    return [
                        'item_id' => $item['item_id'],
                        'sort_order' => $item['sort_order'],
                    ];
                })->all()
            );
        });

        return response()->json([
            'message' => 'updated',
            'outfit' => OutfitPayloadBuilder::buildDetail($outfit->fresh()->load(['outfitItems.item', 'user'])),
        ]);
    });

    Route::middleware('auth:web')->delete('/outfits/{id}', function (Request $request, int $id) {
        $outfit = Outfit::query()
            ->where('user_id', $request->user()->id)
            ->findOrFail($id);

        $outfit->delete();

        return response()->json([
            'message' => 'deleted',
        ]);
    });

    // Wear Logs
    Route::middleware('auth:web')->get('/wear-logs', [WearLogController::class, 'index']);
    Route::middleware('auth:web')->post('/wear-logs', [WearLogController::class, 'store']);
    Route::middleware('auth:web')->get('/wear-logs/calendar', [WearLogController::class, 'calendar']);
    Route::middleware('auth:web')->get('/wear-logs/by-date', [WearLogController::class, 'byDate']);
    Route::middleware('auth:web')->get('/wear-logs/{id}', [WearLogController::class, 'show']);
    Route::middleware('auth:web')->put('/wear-logs/{id}', [WearLogController::class, 'update']);
    Route::middleware('auth:web')->delete('/wear-logs/{id}', [WearLogController::class, 'destroy']);

});

// swagger-ui 表示用
Route::get('/docs/openapi.yaml', function () {
    $path = dirname(base_path()).DIRECTORY_SEPARATOR.'docs'.DIRECTORY_SEPARATOR.'api'.DIRECTORY_SEPARATOR.'openapi.yaml';

    if (! file_exists($path)) {
        abort(404, 'OpenAPI file not found.');
    }

    return response()->file($path, [
        'Content-Type' => 'application/yaml',
    ]);
});