# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Wardrobe App API
  version: 0.1.0
  description: |
    Wardrobe App のAPI仕様。

    このAPIは Cookie ベースのセッション認証を使用する。
    CSRF 保護が必要な POST / PUT / DELETE では X-CSRF-TOKEN ヘッダが必要。
    docs 上では認証系 API / 設定系 API / 参照マスタ API を軸に整理しつつ、Items / Outfits も実装単位で記載する。
    `400` / `429` / `500` は、アプリ独自の契約がある場合を除き framework default に委ね、通常の endpoint 定義では過剰に明示しない。
    `care_status` のような補助状態は、主 status と別カラムで扱い、候補除外や invalid 化の主制御には使わない。
servers:
  - url: http://localhost:8000
    description: ローカル開発環境
...
paths:
...
  /api/register:
    post:
      tags:
        - Auth
      summary: ユーザー登録
      description: |
        新規ユーザーを登録し、成功時はそのままログイン状態にする。
        Cookie ベース認証のため、事前に /csrf-cookie を呼び出し X-CSRF-TOKEN を付与する。
      operationId: register
      parameters:
        - $ref: "#/components/parameters/XCsrfTokenHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: ユーザー登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "422":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
  /api/settings/brands:
    get:
      tags:
        - Settings
      summary: ブランド候補一覧取得
      description: ユーザー単位のブランド候補を返す。active_only と keyword で絞り込みでき、item 新規作成 / 編集のブランド候補サジェストにも利用する。
      security:
        - SessionCookieAuth: []
      parameters:
        - in: query
          name: active_only
          schema:
            type: boolean
          description: true のとき有効な候補だけ返す
        - in: query
          name: keyword
          schema:
            type: string
            maxLength: 255
          description: ブランド名または読み仮名で候補を絞り込む
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBrandsResponse"
        "401":
          description: 未認証
    post:
      tags:
        - Settings
      summary: ブランド候補追加
      description: ブランド候補を user 単位で追加する。重複判定は backend 側で行う。
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/XCsrfTokenHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserBrandUpsertRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBrandResponseWithMessage"
        "401":
          description: 未認証
        "422":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
  /api/settings/brands/{id}:
    patch:
      tags:
        - Settings
      summary: ブランド候補更新
      description: ブランド候補の名称・読み仮名・有効状態を更新する。
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/XCsrfTokenHeader"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserBrandUpdateRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBrandResponseWithMessage"
        "401":
          description: 未認証
        "404":
          description: 未検出（対象なし・他人データ秘匿を含む）
        "422":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
  /api/login:
    post:
      tags:
        - Auth
      summary: ログイン
      description: |
        メールアドレスとパスワードでログインする。
        Cookie ベース認証のため、事前に /csrf-cookie を呼び出し X-CSRF-TOKEN を付与する。
      operationId: login
      parameters:
        - $ref: "#/components/parameters/XCsrfTokenHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "422":
          description: 認証失敗またはバリデーションエラー。認証失敗は message only、入力不備は errors を含む
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/UnprocessableEntityMessageResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
  /api/me:
    get:
      tags:
        - Auth
      summary: ログインユーザー情報取得
      operationId: me
      security:
        - SessionCookieAuth: []
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: 未認証
  /api/logout:
    post:
      tags:
        - Auth
      summary: ログアウト
      description: |
        ルート定義は auth:web 配下に置くが、未認証で /api/logout を呼んだ場合も例外ハンドラ側で logged_out を返す。
        POST のため X-CSRF-TOKEN ヘッダが必要。
      operationId: logout
      parameters:
        - $ref: "#/components/parameters/XCsrfTokenHeader"
      responses:
        "200":
          description: "ログアウト成功。message は `logged_out`、認証状態の正本は Cookie と後続の `/api/me` で確認する"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageOnlyResponse"
...
  /api/settings/categories:
    get:
      tags:
        - Settings
      summary: カテゴリ表示設定取得
      security:
        - SessionCookieAuth: []
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryVisibilitySettingsResponse"
        "401":
          description: 未認証
    put:
      tags:
        - Settings
      summary: カテゴリ表示設定更新
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/XCsrfTokenHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryVisibilitySettingsRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryVisibilitySettingsUpdatedResponse"
        "401":
          description: 未認証
        "422":
          description: バリデーションエラーまたは業務ルール違反。field error がある場合は errors を含み、無効カテゴリなどの業務ルール違反は message only を返す
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ValidationErrorResponse"
                  - $ref: "#/components/schemas/UnprocessableEntityMessageResponse"
  /api/settings/preferences:
    get:
      tags:
        - Settings
      summary: 表示・初期値設定取得
      description: "`currentSeason` と `defaultWearLogStatus` と `calendarWeekStart` と `skinTonePreset` を返す。前者3つの未設定時は null を返し、`skinTonePreset` は `neutral_medium` を返す。"
      security:
        - SessionCookieAuth: []
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferencesResponse"
        "401":
          description: 未認証
    put:
      tags:
        - Settings
      summary: 表示・初期値設定更新
      description: |
        item / outfit 一覧の初期季節と、wear log 新規作成時の初期 status と、wear log カレンダーの週開始と、item サムネイル用の肌色 preset を更新する。
        一覧では URL の季節条件が優先され、URL に季節指定がない場合のみ `currentSeason` を初期値として使う。
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/XCsrfTokenHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserPreferencesRequest"
      responses:
        "200":
          description: "更新成功。message は `updated`、正本は同時に返す preferences と後続の設定取得結果で読む"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferencesUpdatedResponse"
        "401":
          description: 未認証
        "422":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
  /api/settings/tpos:
    get:
      tags:
        - Settings
      summary: TPO 設定一覧取得
      description: |
        現在のユーザーが持つ TPO 一覧を `sort_order asc` で返す。
        `active_only=1` を指定した場合は active な TPO のみ返す。
        settings 管理画面の正本とし、item / outfit の新規候補取得にも使う。
      security:
        - SessionCookieAuth: []
      parameters:
        - name: active_only
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserTposResponse"
        "401":
          description: 未認証
    post:
      tags:
        - Settings
      summary: TPO 追加
      description: |
        ユーザー追加 TPO を新規作成する。
        作成時は `isActive=true`、`isPreset=false`、`sortOrder` は末尾自動採番とする。
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/XCsrfTokenHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserTpoRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserTpoResponseWithMessage"
        "401":
          description: 未認証
        "422":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
...
components:
  parameters:
    XCsrfTokenHeader:
      name: X-CSRF-TOKEN
      in: header
      required: false
      description: |
        CSRF対策用ヘッダ。
        /csrf-cookie 実行後に取得した XSRF-TOKEN と同じ値を指定する。
        実運用では必須。
        Swagger UI では requestInterceptor により自動付与する。
      schema:
        type: string
...
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          maxLength: 50
          example: Test
        email:
          type: string
          format: email
          maxLength: 255
          example: sample.user@gmail.com
        password:
          type: string
          minLength: 8
          maxLength: 72
          example: password123
...
    RegisterResponse:
      type: object
      required:
        - message
        - user
      properties:
        message:
          type: string
          example: registered
        user:
          $ref: "#/components/schemas/User"
...
    ValidationErrorResponse:
      description: |
        field 単位の validation error を返す response。
        `message` は validation 全体の補助説明であり、正本は `errors` を読む。
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: laravel-session